+++ src/hotspot/share/gc/shenandoah/shenandoahControlThread.cpp
@@ -41,6 +41,7 @@
 #include "gc/shenandoah/shenandoahWorkerPolicy.hpp"
 #include "gc/shenandoah/heuristics/shenandoahHeuristics.hpp"
 #include "memory/iterator.hpp"
+#include "memory/metaspaceUtils.hpp"
 #include "memory/universe.hpp"
 #include "runtime/atomic.hpp"
