+++ src/hotspot/share/memory/metaspace/commitLimiter.cpp
@@ -26,6 +26,7 @@
 #include "precompiled.hpp"
 #include "memory/metaspace.hpp"
 #include "memory/metaspace/commitLimiter.hpp"
+#include "memory/metaspaceUtils.hpp"
 #include "utilities/debug.hpp"
 #include "utilities/globalDefinitions.hpp"
