+++ src/hotspot/share/memory/metaspace/metaspaceDCmd.cpp
@@ -27,6 +27,7 @@
 #include "memory/metaspace.hpp"
 #include "memory/metaspace/metaspaceDCmd.hpp"
 #include "memory/metaspace/metaspaceReporter.hpp"
+#include "memory/metaspaceUtils.hpp"
 #include "memory/resourceArea.hpp"
 #include "runtime/vmOperations.hpp"
 #include "services/diagnosticCommand.hpp"
