+++ src/hotspot/share/memory/metaspaceTracer.hpp
@@ -27,6 +27,7 @@
 
 #include "memory/allocation.hpp"
 #include "memory/metaspace.hpp"
+#include "memory/metaspaceUtils.hpp"
 
 class ClassLoaderData;
