+++ src/hotspot/share/oops/metadata.cpp
@@ -24,7 +24,6 @@
 
 #include "precompiled.hpp"
 #include "oops/metadata.hpp"
-#include "memory/metaspace.hpp"
 #include "memory/resourceArea.hpp"
 #include "prims/jvmtiRedefineClasses.hpp"
