+++ test/hotspot/gtest/metaspace/test_metaspaceUtils.cpp
@@ -25,6 +25,7 @@
 
 #include "precompiled.hpp"
 #include "memory/metaspace.hpp"
+#include "memory/metaspaceUtils.hpp"
 #include "unittest.hpp"
 
 TEST_VM(MetaspaceUtils, reserved) {
